strncat

+++ 0x06-pointers_arrays_strings/1-strncat.c
@@ -12,7 +12,7 @@
 char *_strncat(char *dest, char *src, int n)
 {
     size_t destlen = strlen(src);
-    int x;
+    size_t x;
 
     for (x = 0; x < n && src[x] != '\0'; x++)
     {
